# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https:///www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://kids-collage.test/api",
            "description": "Local Host"
        }
    ],
    "paths": {
...
        "/splash": {
            "get": {
                "tags": [
                    "Splash"
                ],
                "summary": "splash",
                "description": "Returns splash",
                "operationId": "splashAdmin",
                "parameters": [
                    {
                        "name": "browser",
                        "in": "query",
                        "description": "Browser name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "chrome",
                            "enum": [
                                "chrome",
                                "firefox",
                                "safari",
                                "edge",
                                "opera",
                                "ie",
                                "other"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "example message"
                                        },
                                        "data": {
                                            "properties": {
                                                "version": {
                                                    "type": "string",
                                                    "default": "1.0.0"
                                                },
                                                "supported_locales": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "label": {
                                                                "type": "string",
                                                                "default": "English"
                                                            },
                                                            "value": {
                                                                "type": "string",
                                                                "default": "en"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "user": {
                                                    "$ref": "#/components/schemas/UserDetailResource"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "default": "Admin"
                                                    }
                                                },
                                                "permissions": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "default": "Shared.Admin"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserDetailResource": {
                "title": "UserDetailResource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "default": "1"
                    },
                    "title": {
                        "type": "string",
                        "default": "User Title"
                    },
                    "description": {
                        "type": "string",
                        "default": "User Description"
                    },
                    "updated_at": {
                        "type": "string",
                        "default": "2024-08-19T07:26:07.000000Z"
                    },
                    "created_at": {
                        "type": "string",
                        "default": "2024-08-19T07:26:07.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            },
            "clubId": {
                "type": "apiKey",
                "description": "Club ID header",
                "name": "X-Club-Id",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": [],
            "clubId": []
        }
    ]
}